refactor: ListBasedReservationDao를 테스트 패키지로 이동
- 프로덕션에서는 사용하지 않고 테스트에서만 Fake 객체로 사용되기 때문에 테스트 패키지로 이동했다.
클래스명도 FakeReservationDao로 수정하였다.
테스트에서는 동시성 문제를 고려하지 않아도 되므로 AtomicLong이 아닌 long으로 index를 관리한다.
